backend/requirements.txt
# ============================================
# ECOMMERCE BACKEND - PYTHON DEPENDENCIES
# ============================================
#
# Python Version Support: 3.10, 3.11, 3.12, 3.13
# WARNING: Python 3.14+ may have compatibility issues with some packages
#
# Last Updated: 2025-10-16
# ============================================

# ==================
# CORE FRAMEWORK
# ==================
fastapi==0.119.0
uvicorn[standard]==0.37.0
bleach

# ==================
# DATABASE & ORM
# ==================
sqlmodel==0.0.27
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# ==================
# AUTHENTICATION & SECURITY
# ==================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ==================
# CONFIGURATION
# ==================
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# ==================
# PAYMENT PROCESSING
# ==================
stripe==11.3.0

# ==================
# UTILITIES
# ==================
aiofiles==24.1.0

# ==================
# DEVELOPMENT / OPTIONAL
# ==================
# Uncomment for placeholder generation
Pillow==11.0.0
Faker==33.1.0

# ==================
# TESTING (Optional)
# ==================
# pytest==8.3.4
# pytest-asyncio==0.24.0
# httpx==0.28.1

# ==================
# PRODUCTION (Optional)
# ==================
# gunicorn==23.0.0
# redis==5.2.1
# celery==5.4.0
